[FLINK-10715] Suppress ConcurrnetModificationException in Slf4jReporter
This commit suppress the ConcurrentModificationException thrown by the Slf4jReporter
report method. This exception occurs when a metric is added/removed while this report is reporting.
Suppressing this (somewhat expected exception) is needed by our e2e framework to reduce some
false-positive e2e test failures. This does not solves the root issue, which is being tracked in a different issue FLINK-10035.

This closes #7008.